Enhancements for REST controllers.
