+++ src/main/java/org/craftercms/core/controller/rest/RestControllerBase.java
@@ -35,74 +35,20 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 /**
- * Base class for all Crafter REST services. Handles exceptions thrown by service methods.
+ * Base class for Crafter REST services
  *
- * @author Alfonso Vásquez
+ * @author avasquez
  */
 public class RestControllerBase {
-
-    private static final Log logger = LogFactory.getLog(RestControllerBase.class);
 
     public static final String REST_BASE_URI = "${crafter.core.rest.base.uri}";
     public static final String MESSAGE_MODEL_ATTRIBUTE_NAME = "message";
 
-    @ExceptionHandler(InvalidContextException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ResponseBody
-    public Map<String, Object> handleInvalidContextException(HttpServletRequest request, InvalidContextException e) {
-        return handleException(request, e);
+    protected Map<String, Object> createResponseMessage(String message) {
+        return createSingletonModifiableMap(MESSAGE_MODEL_ATTRIBUTE_NAME, message);
     }
 
-    @ExceptionHandler(AuthenticationException.class)
-    @ResponseStatus(HttpStatus.UNAUTHORIZED)
-    @ResponseBody
-    public Map<String, Object> handleAuthenticationException(HttpServletRequest request, AuthenticationException e) {
-        return handleException(request, e);
-    }
-
-    @ExceptionHandler(PathNotFoundException.class)
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    @ResponseBody
-    public Map<String, Object> handlePathNotFoundException(HttpServletRequest request, PathNotFoundException e) {
-        return handleException(request, e);
-    }
-
-    @ExceptionHandler(ForbiddenPathException.class)
-    @ResponseStatus(HttpStatus.FORBIDDEN)
-    @ResponseBody
-    public Map<String, Object> handleForbiddenPathException(HttpServletRequest request, ForbiddenPathException e) {
-        return handleException(request, e);
-    }
-
-    @ExceptionHandler(ValidationException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ResponseBody
-    public ValidationResult handleValidationException(HttpServletRequest request, ValidationException e) {
-        logger.error("Request for " + request.getRequestURI() + " failed", e);
-
-        return e.getResult();
-    }
-
-    @ExceptionHandler(ValidationRuntimeException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ResponseBody
-    public ValidationResult handleValidationRuntimeException(HttpServletRequest request, ValidationRuntimeException e) {
-        logger.error("Request for " + request.getRequestURI() + " failed", e);
-
-        return e.getResult();
-    }
-
-    @ExceptionHandler(Exception.class)
-    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
-    @ResponseBody
-    public Map<String, Object> handleException(HttpServletRequest request, Exception e) {
-        logger.error("Request for " + request.getRequestURI() + " failed", e);
-
-        return createMessageModel(MESSAGE_MODEL_ATTRIBUTE_NAME, e.getMessage());
-    }
-
-    // Use instead of singleton model since it can modified
-    protected Map<String, Object> createMessageModel(String attributeName, Object attributeValue) {
+    protected Map<String, Object> createSingletonModifiableMap(String attributeName, Object attributeValue) {
         Map<String, Object> model = new HashMap<>(1);
         model.put(attributeName, attributeValue);
